[mbr] Fixup console sample
Use a single changing testfile
