+++ src/mono/netcore/sample/mbr/console/Program.cs
@@ -37,11 +37,3 @@ namespace HelloWorld
     }
 }
 
-public class TestClass {
-	[MethodImpl(MethodImplOptions.NoInlining)]
-	public static string TargetMethod () {
-		string s = "OLD STRING";
-		Console.WriteLine (s);
-		return s;
-        }
-}
